removed the dependency on stdweb - resolves #144

+++ examples/svg.rs
@@ -11,8 +11,8 @@ fn main() {
     let current_layer = doc.get_page(page1).get_layer(layer1);
     let svg = Svg::parse(SVG).unwrap();
 
-    let rotation_center_x = Px((svg.width.0 as f64 / 2.0) as usize);
-    let rotation_center_y = Px((svg.height.0 as f64 / 2.0) as usize);
+    let rotation_center_x = Px((svg.width.0 as f32 / 2.0) as usize);
+    let rotation_center_y = Px((svg.height.0 as f32 / 2.0) as usize);
 
     let reference = svg.into_xobject(&current_layer);
 
@@ -21,12 +21,12 @@ fn main() {
             &current_layer,
             SvgTransform {
                 rotate: Some(SvgRotation {
-                    angle_ccw_degrees: i as f64 * 36.0,
+                    angle_ccw_degrees: i as f32 * 36.0,
                     rotation_center_x: rotation_center_x.into_pt(300.0),
                     rotation_center_y: rotation_center_y.into_pt(300.0),
                 }),
-                translate_x: Some(Mm(i as f64 * 20.0 % 50.0).into()),
-                translate_y: Some(Mm(i as f64 * 30.0).into()),
+                translate_x: Some(Mm(i as f32 * 20.0 % 50.0).into()),
+                translate_y: Some(Mm(i as f32 * 30.0).into()),
                 ..Default::default()
             },
         );
